+++ src/buffer/mapshaper-buffer-common.js
@@ -1,4 +1,3 @@
-
 import { compileValueExpression } from '../expressions/mapshaper-expressions';
 import { getDatasetCRS } from '../crs/mapshaper-projections';
 import { convertDistanceParam } from '../geom/mapshaper-units';